+++ Gemfile
@@ -1,2 +1,3 @@
 gem 'rubocop'
+gem 'rspec'
 source 'https://rubygems.org'
