+++ bin/main.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 require_relative '../lib/game_logic.rb'
-require_relative '../lib/player.rb'
+require_relative '../lib/players.rb'
 
 def display_board
   @field.each_index do |i|
@@ -44,30 +44,6 @@ def game_over
   end
 end
 
-def check_win(player, player_history_sorted)
-  WIN.each do |r1|
-    counter = 0
-
-    r1.each do |r2|
-      counter += 1 if player_history_sorted.include?(r2.to_s)
-
-      next unless counter == 3
-
-      display_board
-      puts "Wow, #{player.name}! It's a win!"
-      game_over
-    end
-  end
-
-  if @turn_count >= 9
-    display_board
-    puts "It's a tie.Try again?"
-    game_over
-  end
-
-  turn
-end
-
 def start_game
   print "Lets start\n"
   p1_name = nil
@@ -86,9 +62,6 @@ def start_game
     break unless p2_name.empty?
   end
 
-  p1 = Player.new(p1_name, 'X')
-  p2 = Player.new(p2_name, 'O')
-  @game = Game.new(p1, p2)
-
-  @game.turn
+  res = [p1_name, p2_name]
 end
+
