+++ lib/game_logic.rb
@@ -20,5 +20,37 @@ class Game
   end
 end
 
+def check_win(player, player_history_sorted)
+  WIN.each do |r1|
+    counter = 0
+
+    r1.each do |r2|
+      counter += 1 if player_history_sorted.include?(r2.to_s)
+
+      next unless counter == 3
+
+      display_board
+      puts "Wow, #{player.name}! It's a win!"
+      game_over
+    end
+  end
+
+  if @turn_count >= 9
+    display_board
+    puts "It's a tie.Try again?"
+    game_over
+  end
+
+  turn
+end
+
 WIN = [[1, 2, 3], [4, 5, 6], [7, 8, 9], [1, 4, 7], [2, 5, 8], [3, 6, 9], [1, 5, 9], [3, 5, 7]].freeze
-start_game
+res =  start_game
+
+p1 = Player.new(res[0], 'X')
+  p2 = Player.new(res[1], 'O')
+  puts p1
+  puts p2
+  games = Game.new(p1, p2)
+
+  games.turn
